bbbb

+++ src/components/BookCard.jsx
@@ -1,24 +1,39 @@
+import { useCallback } from "react";
 import { Link, useLocation } from "react-router-dom";
 import { deleteBook } from "../libs/services/slices/booksSlice";
 import { useDispatch, useSelector } from "react-redux";
+import { showSuccessToast, showErrorToast } from "../libs/toastNotifications";
+
 
 export default function BookCard({ book }) {
   const user = useSelector((state) => state.auth.user);
+  const error = useSelector((state) => state.books.error); 
   const dispatch = useDispatch();
   const location = useLocation();
   const isAdminPage = location.pathname === "/admin";
-  const handleDelete = (id, title) => {
-    const confirmDelete = window.confirm(`هل تريد حقًا حذف كتاب  "${title}"؟`);
+
+  const handleDelete = useCallback(
+    async (id, title) => {
+      try {
+ const confirmDelete = window.confirm(`هل تريد حقًا حذف كتاب  "${title}"؟`);
     if (confirmDelete) {
       dispatch(deleteBook(id));
+showSuccessToast("تم حذف الكتاب بنجاح");
     }
-  };
+      } catch (err) {
+        showErrorToast("حدث خطأ أثناء حذف الكتاب");
+      }
+    },
+    []
+  );
+
+
 
   return (
     <div className="border rounded-lg shadow bg-white text-text">
       <div className="h-48 w-full bg-gray-300 text-center flex items-center justify-center">
         <img
-          src="../assits/images/books.gif"
+          src="../assits/images/books.png"
           width={150}
           height={150}
           alt="book"
@@ -29,12 +44,12 @@ export default function BookCard({ book }) {
           {book?.title}
         </h2>
         <div className="w-full px-2 flex justify-between items-center">
-          <p className="text-sm text-gray-700 mb-1">{book?.author}</p>
-          <p className="text-sm text-gray-700">{book?.category}</p>
+           <p className="text-sm text-gray-700">{book?.category}</p>
+<p className="text-sm text-gray-700 mb-1">{book?.author}</p>
         </div>
       </div>
       {user && user.role === "admin" && isAdminPage && (
-        <div className="w-full flex justify-around items-center p-2">
+        <div className="w-full flex justify-around items-center p-2 py-4">
           <Link
             to={`/admin/edit/${book?.id}`}
             className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded"
